requirements.txt
pydantic
pydantic_ai
streamlit
pypdf2
python-dotenv
groq
requests
rich 
pymupdf
beautifulsoup4
lxml
chromadb
sentence-transformers